fix: split server fns from db-heavy helpers to fix client bundle
The overview page's <Lazy> component error was caused by
prompt-wizard.tsx importing from @/server/onboarding, which mixed
createServerFn server functions with regular exports that transitively
pull in db/drizzle/pg. TanStack Start replaces server functions with
RPC stubs on the client, but only if the module doesn't also export
server-only code that gets bundled.

The old @/server/wizard.ts (pre-PR) only exported server functions —
that's why it worked. The new onboarding.ts added createBrand,
updateBrand, buildBrandResult, BrandConflictError, etc. alongside the
server fns, breaking the client-side tree-shaking.

Split into two files:
- onboarding.ts: server functions only (analyzeBrandFn,
  updateOnboardedBrandFn). No db imports. Client components import
  from here.
- onboarding-core.ts: everything else (createBrand, updateBrand,
  schemas, error classes, helpers). API routes import from here.
  Server-only.
