+++ apps/web/src/server/onboarding.ts
@@ -1,376 +1,21 @@
 /**
- * Brand-onboarding server functions + the shared persistence helpers used by
- * both the in-app wizard and the public REST endpoints under /api/v1/brands.
+ * Server functions for the onboarding wizard + brand analysis.
  *
- * Three exported building blocks:
- *   • createBrand — pure create, no LLM, no merging. Throws BrandConflictError.
- *   • updateBrand — pure brand-level update (replace semantics on arrays).
- *     Throws BrandNotFoundError.
- *   • updateOnboardedBrandFn — server fn used only by the in-app wizard's save
- *     step. Bundles brand-field replace + net-new prompt/competitor inserts
- *     (deduped against existing DB rows) into one round-trip. Not exposed via
- *     the public API.
+ * This file ONLY exports createServerFn server functions. No regular function
+ * exports, no class exports, no db imports. TanStack Start replaces server
+ * functions with lightweight RPC stubs on the client — but only if the module
+ * doesn't drag in server-only dependencies (db, drizzle, pg) via other
+ * exports. Keeping this file server-fn-only guarantees the client bundle
+ * stays clean.
  *
- * The public API never invokes analyzeBrand under the hood. Callers who want
- * suggestions hit POST /api/v1/tools/analyze first, then feed the result into
- * POST /api/v1/brands themselves.
+ * Regular functions (createBrand, updateBrand, etc.) live in
+ * ./onboarding-core.ts, imported only by API routes (server-only).
  */
 import { createServerFn } from "@tanstack/react-start";
 import { z } from "zod";
-import { eq, count } from "drizzle-orm";
-import { db } from "@workspace/lib/db/db";
-import { brands, prompts, competitors } from "@workspace/lib/db/schema";
-import { MAX_COMPETITORS } from "@workspace/lib/constants";
-import { computeSystemTags, sanitizeUserTags } from "@workspace/lib/tag-utils";
 import { analyzeBrand } from "@workspace/lib/onboarding";
 import { requireAuthSession, requireOrgAccess } from "@/lib/auth/helpers";
-import { dedupeDomains, dedupeAliases } from "@/lib/domain-categories";
-import { createMultiplePromptJobSchedulers } from "@/lib/job-scheduler";
-
-// ============================================================================
-// Errors
-// ============================================================================
-
-export class BrandConflictError extends Error {
-	constructor(public readonly brandId: string) {
-		super(`Brand "${brandId}" already exists.`);
-		this.name = "BrandConflictError";
-	}
-}
-
-export class BrandNotFoundError extends Error {
-	constructor(public readonly brandId: string) {
-		super(`Brand "${brandId}" not found.`);
-		this.name = "BrandNotFoundError";
-	}
-}
-
-// ============================================================================
-// Schemas
-// ============================================================================
-
-const competitorInputSchema = z.object({
-	name: z.string().min(1),
-	domains: z.array(z.string()).optional().default([]),
-	aliases: z.array(z.string()).optional().default([]),
-});
-
-const promptInputSchema = z.object({
-	value: z.string().min(1),
-	tags: z.array(z.string()).optional().default([]),
-	enabled: z.boolean().optional().default(true),
-});
-
-/** POST /api/v1/brands body. */
-const createBrandInputSchema = z.object({
-	id: z.string().min(1),
-	name: z.string().min(1),
-	website: z.string().min(1),
-	additionalDomains: z.array(z.string()).optional(),
-	aliases: z.array(z.string()).optional(),
-	competitors: z.array(competitorInputSchema).optional(),
-	prompts: z.array(promptInputSchema).optional(),
-});
-
-/** PATCH /api/v1/brands/:brandId body. brandId comes from the URL. */
-const updateBrandBodySchema = z.object({
-	brandName: z.string().min(1).optional(),
-	website: z.string().min(1).optional(),
-	additionalDomains: z.array(z.string()).optional(),
-	aliases: z.array(z.string()).optional(),
-	enabled: z.boolean().optional(),
-});
-
-const updateBrandInputSchema = updateBrandBodySchema.extend({
-	brandId: z.string().min(1),
-});
-
-/** Wizard save: brand-level fields + new prompts/competitors in one shot. */
-const wizardOnboardingInputSchema = z.object({
-	brandId: z.string().min(1),
-	brandName: z.string().min(1).optional(),
-	website: z.string().min(1).optional(),
-	additionalDomains: z.array(z.string()).optional(),
-	aliases: z.array(z.string()).optional(),
-	competitors: z.array(competitorInputSchema).optional(),
-	prompts: z.array(promptInputSchema).optional(),
-});
-
-export type CreateBrandInput = z.infer<typeof createBrandInputSchema>;
-export type UpdateBrandInput = z.infer<typeof updateBrandInputSchema>;
-export type WizardOnboardingInput = z.infer<typeof wizardOnboardingInputSchema>;
-
-export interface BrandResult {
-	id: string;
-	name: string;
-	website: string;
-	additionalDomains: string[];
-	aliases: string[];
-	enabled: boolean;
-	onboarded: boolean;
-	createdAt: Date;
-	updatedAt: Date;
-}
-
-// ============================================================================
-// Helpers
-// ============================================================================
-
-function validateAndFormatWebsite(url: string): string {
-	const trimmed = url.trim();
-	const formatted = trimmed.startsWith("http://") || trimmed.startsWith("https://") ? trimmed : `https://${trimmed}`;
-	const parsed = new URL(formatted);
-	if (!["http:", "https:"].includes(parsed.protocol)) {
-		throw new Error("Website URL must use http or https");
-	}
-	if (!parsed.hostname) {
-		throw new Error("Website URL must have a valid hostname");
-	}
-	return formatted;
-}
-
-export function buildBrandResult(row: typeof brands.$inferSelect): BrandResult {
-	return {
-		id: row.id,
-		name: row.name,
-		website: row.website,
-		additionalDomains: row.additionalDomains,
-		aliases: row.aliases,
-		enabled: row.enabled,
-		onboarded: row.onboarded,
-		createdAt: row.createdAt,
-		updatedAt: row.updatedAt,
-	};
-}
-
-async function insertCompetitors(args: {
-	brandId: string;
-	websiteHost: string;
-	source: { name: string; domains: string[]; aliases: string[] }[];
-}): Promise<number> {
-	if (args.source.length === 0) return 0;
-
-	const existing = await db.query.competitors.findMany({
-		where: eq(competitors.brandId, args.brandId),
-	});
-	const existingDomains = new Set(existing.flatMap((c) => c.domains));
-
-	const toInsert: Array<{ brandId: string; name: string; domains: string[]; aliases: string[] }> = [];
-	for (const c of args.source) {
-		const cleaned = dedupeDomains(c.domains).filter((d) => d !== args.websiteHost);
-		if (cleaned.length === 0) continue;
-		if (cleaned.some((d) => existingDomains.has(d))) continue;
-		toInsert.push({
-			brandId: args.brandId,
-			name: c.name.trim(),
-			domains: cleaned,
-			aliases: dedupeAliases(c.aliases),
-		});
-	}
-	if (toInsert.length === 0) return 0;
-
-	const [{ count: currentCount }] = await db
-		.select({ count: count() })
-		.from(competitors)
-		.where(eq(competitors.brandId, args.brandId));
-	if ((currentCount || 0) + toInsert.length > MAX_COMPETITORS) {
-		throw new Error(
-			`Cannot add competitors. Would exceed maximum of ${MAX_COMPETITORS} (currently ${currentCount}, adding ${toInsert.length}).`,
-		);
-	}
-
-	await db.insert(competitors).values(toInsert);
-	return toInsert.length;
-}
-
-async function insertPrompts(args: {
-	brandId: string;
-	brandName: string;
-	website: string;
-	source: { value: string; tags: string[]; enabled: boolean }[];
-	dedupeAgainstExisting: boolean;
-}): Promise<number> {
-	if (args.source.length === 0) return 0;
-
-	const seen = new Set<string>();
-	if (args.dedupeAgainstExisting) {
-		const existing = await db.query.prompts.findMany({
-			where: eq(prompts.brandId, args.brandId),
-		});
-		for (const p of existing) seen.add(p.value.toLowerCase());
-	}
-
-	const rows: Array<{
-		brandId: string;
-		value: string;
-		enabled: boolean;
-		tags: string[];
-		systemTags: string[];
-	}> = [];
-	for (const p of args.source) {
-		const value = p.value.trim();
-		if (!value) continue;
-		const key = value.toLowerCase();
-		if (seen.has(key)) continue;
-		seen.add(key);
-		rows.push({
-			brandId: args.brandId,
-			value,
-			enabled: p.enabled,
-			tags: p.tags,
-			systemTags: computeSystemTags(value, args.brandName, args.website),
-		});
-	}
-	if (rows.length === 0) return 0;
-
-	const inserted = await db.insert(prompts).values(rows).returning({ id: prompts.id });
-	await createMultiplePromptJobSchedulers(inserted.map((r) => r.id));
-	return inserted.length;
-}
-
-// ============================================================================
-// createBrand — pure create
-// ============================================================================
-
-/**
- * Create a brand row (and optional initial competitors/prompts) exactly as
- * the caller supplied them. No LLM call, no merging with anything else.
- * Throws BrandConflictError if a brand with this brandId already exists.
- */
-export async function createBrand(input: CreateBrandInput): Promise<BrandResult> {
-	const formattedWebsite = validateAndFormatWebsite(input.website);
-	const websiteHost = new URL(formattedWebsite).hostname.replace(/^www\./, "");
-
-	const additionalDomains = dedupeDomains(input.additionalDomains ?? []).filter((d) => d !== websiteHost);
-	const aliases = dedupeAliases(input.aliases ?? []);
-
-	const [inserted] = await db
-		.insert(brands)
-		.values({
-			id: input.id,
-			name: input.name,
-			website: formattedWebsite,
-			additionalDomains,
-			aliases,
-			enabled: true,
-			onboarded: true,
-		})
-		.onConflictDoNothing()
-		.returning({ id: brands.id });
-	if (!inserted) throw new BrandConflictError(input.id);
-
-	await insertCompetitors({
-		brandId: input.id,
-		websiteHost,
-		source: (input.competitors ?? []).map((c) => ({
-			name: c.name,
-			domains: c.domains ?? [],
-			aliases: c.aliases ?? [],
-		})),
-	});
-
-	await insertPrompts({
-		brandId: input.id,
-		brandName: input.name,
-		website: formattedWebsite,
-		source: (input.prompts ?? []).map((p) => ({
-			value: p.value,
-			tags: sanitizeUserTags(p.tags ?? []),
-			enabled: p.enabled ?? true,
-		})),
-		// Brand row was just created — no existing prompts to dedupe against.
-		dedupeAgainstExisting: false,
-	});
-
-	const refreshed = await db.query.brands.findFirst({ where: eq(brands.id, input.id) });
-	return buildBrandResult(refreshed!);
-}
-
-// ============================================================================
-// updateBrand — pure brand-level update
-// ============================================================================
-
-/**
- * Update brand-level fields. Replace semantics: any array field the caller
- * provides replaces the existing value verbatim. Throws BrandNotFoundError
- * if the brand doesn't exist. Doesn't touch prompts or competitors — those
- * are managed via /api/v1/prompts and /api/v1/competitors.
- */
-export async function updateBrand(input: UpdateBrandInput): Promise<BrandResult> {
-	const existing = await db.query.brands.findFirst({ where: eq(brands.id, input.brandId) });
-	if (!existing) throw new BrandNotFoundError(input.brandId);
-
-	const formattedWebsite = input.website ? validateAndFormatWebsite(input.website) : null;
-	const websiteHost = formattedWebsite
-		? new URL(formattedWebsite).hostname.replace(/^www\./, "")
-		: existing.website
-			? new URL(existing.website).hostname.replace(/^www\./, "")
-			: null;
-
-	const patch: Partial<typeof brands.$inferInsert> = { updatedAt: new Date() };
-	if (input.brandName !== undefined) patch.name = input.brandName;
-	if (formattedWebsite !== null) patch.website = formattedWebsite;
-	if (input.additionalDomains !== undefined) {
-		patch.additionalDomains = dedupeDomains(input.additionalDomains).filter((d) => d !== websiteHost);
-	}
-	if (input.aliases !== undefined) patch.aliases = dedupeAliases(input.aliases);
-	if (input.enabled !== undefined) patch.enabled = input.enabled;
-
-	await db.update(brands).set(patch).where(eq(brands.id, input.brandId));
-	const refreshed = await db.query.brands.findFirst({ where: eq(brands.id, input.brandId) });
-	return buildBrandResult(refreshed!);
-}
-
-// ============================================================================
-// Wizard save — brand fields + new prompts/competitors in one shot
-// ============================================================================
-
-async function saveWizardOnboarding(input: WizardOnboardingInput): Promise<BrandResult> {
-	// Use updateBrand for the brand-level fields (replace semantics)
-	await updateBrand({
-		brandId: input.brandId,
-		brandName: input.brandName,
-		website: input.website,
-		additionalDomains: input.additionalDomains,
-		aliases: input.aliases,
-	});
-
-	// Also set onboarded flag (updateBrand doesn't do this)
-	await db.update(brands).set({ onboarded: true, updatedAt: new Date() }).where(eq(brands.id, input.brandId));
-
-	const existing = await db.query.brands.findFirst({ where: eq(brands.id, input.brandId) });
-	if (!existing) throw new BrandNotFoundError(input.brandId);
-	const websiteHost = new URL(existing.website).hostname.replace(/^www\./, "");
-
-	await insertCompetitors({
-		brandId: input.brandId,
-		websiteHost,
-		source: (input.competitors ?? []).map((c) => ({
-			name: c.name,
-			domains: c.domains ?? [],
-			aliases: c.aliases ?? [],
-		})),
-	});
-
-	await insertPrompts({
-		brandId: input.brandId,
-		brandName: existing.name,
-		website: existing.website,
-		source: (input.prompts ?? []).map((p) => ({
-			value: p.value,
-			tags: sanitizeUserTags(p.tags ?? []),
-			enabled: p.enabled ?? true,
-		})),
-		dedupeAgainstExisting: true,
-	});
-
-	const refreshed = await db.query.brands.findFirst({ where: eq(brands.id, input.brandId) });
-	return buildBrandResult(refreshed!);
-}
-
-// ============================================================================
-// Server functions (in-app wizard)
-// ============================================================================
+import { saveWizardOnboarding, wizardOnboardingInputSchema } from "@/server/onboarding-core";
 
 /** Run brand analysis without saving anything. */
 export const analyzeBrandFn = createServerFn({ method: "POST" })
@@ -394,9 +39,7 @@ export const analyzeBrandFn = createServerFn({ method: "POST" })
 
 /**
  * Persist the wizard's reviewed onboarding result for a brand the user
- * already has access to. The brand row was created earlier in the signup
- * flow (see BrandOnboarding); this just tops it up with the user-confirmed
- * brand fields, prompts, and competitors.
+ * already has access to.
  */
 export const updateOnboardedBrandFn = createServerFn({ method: "POST" })
 	.inputValidator(wizardOnboardingInputSchema)
@@ -405,8 +48,3 @@ export const updateOnboardedBrandFn = createServerFn({ method: "POST" })
 		await requireOrgAccess(session.user.id, data.brandId);
 		return saveWizardOnboarding(data);
 	});
-
-export {
-	createBrandInputSchema,
-	updateBrandBodySchema,
-};
